Move raise condition for host into enabled method
This will fix the ssl resource until #1205 is rolled out in full. Still
looking around the code and need to build some unit tests for the ssl
resource.

My fix here is to move the raise condition till later in the flow,
specifically the enabled? method. This lets the raise get caught
accordingly without killing the other tests.

Signed-off-by: Jared Quick <jquick@chef.io>